+++ bin/start_varlogmr.py
@@ -1,13 +1,16 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
+import argparse
 import glob
 import json
 import os
+import shutil
 import socket
 import subprocess
 import sys
 import time
+from typing import List
 
 cwd = os.path.dirname(os.path.realpath(__file__))  # noqa
 binpath = os.path.join(cwd, "..", "bin")  # noqa
@@ -15,19 +18,19 @@ pylib = os.path.join(cwd, "..", "pylib")  # noqa
 if os.path.isdir(pylib):  # noqa
     sys.path.insert(0, pylib)  # noqa
 
-from varlog.killer import Killer  # noqa
-from varlog import procutil  # noqa
 from varlog import limits  # noqa
+from varlog import procutil  # noqa
+from varlog.killer import Killer  # noqa
 from varlog.logger import get_logger  # noqa
 
-logger = get_logger("vmr")
-
+APP_NAME = "vmr"
 DEFAULT_CLUSTER_ID = 1
-RETRY_INTERVAL_SEC = 3
+DEFAULT_RPC_PORT = 9092
+DEFAULT_RAFT_PORT = 10000
+DEFAULT_RETRY_INTERVAL_SECONDS = 3
+DEFAULT_ADD_AFTER_SECONDS = 5
 
-DEFAULT_REP_FACTOR = 1
-DEFAULT_VMR_HOME = '/home/deploy/varlog-mr'
-LOCAL_ADDRESS = socket.gethostbyname(socket.gethostname())
+logger = get_logger(APP_NAME)
 
 
 class Node:
@@ -37,124 +40,54 @@ class Node:
         raft_address: An address to use RAFT protocol
         rpc_address: An address to use RPC communications.
     """
-
     def __init__(self, raft_address: str, rpc_address: str):
         self.raft_address = raft_address
         self.rpc_address = rpc_address
 
 
-def get_cluster_id() -> int:
-    """Get Cluster ID.
-
-    Returns:
-        Cluster ID.
-    """
-    return int(os.getenv('CLUSTER_ID', str(DEFAULT_CLUSTER_ID)))
-
-
-def get_raft_address() -> str:
-    """Get local RAFT address.
-
-    Get RAFT address defined by environment variable RAFT_ADDRESS.
-
-    Raises:
-        ValueError: An error occurred when not defined RAFT_ADDRESS.
-    """
-    raft_addr = os.getenv('RAFT_ADDRESS')
-    if raft_addr is None:
-        raise ValueError('no raft address')
-    return raft_addr
-
-
-def get_rpc_address() -> str:
-    """Get local RPC address.
-
-    Get local RPC address defined by environment variable RPC_ADDRESS.
-
-    Raises:
-        ValueError: An error occurred when not defined RPC_ADDRESS.
-    """
-    rpc_addr = os.getenv('RPC_ADDRESS')
-    if rpc_addr is None:
-        raise ValueError('no rpc address')
-    return rpc_addr
-
-
-def get_seed_address() -> str:
-    """Get seed metadata repository address.
-
-    Get seed metadata repository address defined by environment variable
-    SEED_ADDRESS. If the environment variable is not defined, empty string is
-    returned.
-    """
-    return os.getenv('SEED_ADDRESS', '')
-
-
-def get_vms_addr() -> str:
-    addr = os.getenv('VMS_ADDRESS')
-    if addr is None:
-        raise ValueError('no admin address')
-    return addr
-
-
-def get_raft_dir() -> str:
-    home = os.getenv('VMR_HOME', DEFAULT_VMR_HOME)
-    return f"{home}/raftdata"
-
-
-def get_log_dir() -> str:
-    home = os.getenv('VMR_HOME', DEFAULT_VMR_HOME)
-    return f"{home}/log"
-
-
-def get_replication_factor() -> int:
-    return int(os.getenv('REPLICATION_FACTOR', str(DEFAULT_REP_FACTOR)))
-
-
-def add_raft_peer() -> bool:
+def add_raft_peer(admin: str, raft_addr: str, rpc_addr: str) -> bool:
     """Add this node to metadata repository cluster.
 
     Returns:
         True if this node is added to the metadata repository cluster
         successfully.
-
-    Raises:
-        ValueError: If environment variables RAFT_ADDRESS and RPC_ADDRESS are
-          not defined, it raises ValueError.
-        subprocess.CalledProcessError:
     """
-    raft_url = get_raft_address()
-    rpc_addr = get_rpc_address()
-    cmd = [
-        f'{binpath}/varlogctl',
-        'mr',
-        'add',
-        f'--raft-url={raft_url}',
-        f'--rpc-addr={rpc_addr}',
-        f'--admin={get_vms_addr()}'
-    ]
-    logger.info(f'add peer: {cmd}')
-    out = subprocess.check_output(cmd)
-    result = json.loads(out)
-    data = result.get('data', dict())
-    items = data.get('items', list())
-    node_id = items[0].get('nodeId', 0)
-    return node_id != 0
-
-
-def prepare_vmr_home():
-    home = os.getenv("VMR_HOME", DEFAULT_VMR_HOME)
-    os.makedirs(home, exist_ok=True)
+    try:
+        cmd = [
+            f"{binpath}/varlogctl",
+            "mr",
+            "add",
+            f"--raft-url={raft_addr}",
+            f"--rpc-addr={rpc_addr}",
+            f"--admin={admin}",
+        ]
+        logger.info(f"add_raft_peer: {cmd}")
+        out = subprocess.check_output(cmd)
+        result = json.loads(out)
+        data = result.get("data", dict())
+        items = data.get("items", list())
+        node_id = items[0].get("nodeId", 0)
+        return node_id != 0
+    except subprocess.CalledProcessError:
+        return False
 
 
 def exists_wal(path: str) -> bool:
-    wal = glob.glob(os.path.join(f'{path}/wal', '*', '*.wal'))
+    """Check if the WAL file exists.
+
+    Args:
+        path: Path to the WAL file.
+
+    Returns:
+        True if the WAL file exists.
+    """
+    wal = glob.glob(os.path.join(f"{path}/wal", "*", "*.wal"))
     for f in wal:
-        logger.info(f'wal: {f}')
+        logger.info(f"WAL: {f}")
     return len(wal) > 0
 
 
-def fetch_peers(cluster_id: int, seed_addr: str) -> list[Node]:
+def fetch_peers(cluster_id: int, seed_addr: str) -> List[Node]:
     """Get a list of Nodes from a seed metadata repository.
 
     Fetch a list of Nodes by connecting to the metadata repository identified
@@ -168,64 +101,80 @@ def fetch_peers(cluster_id: int, seed_addr: str) -> list[Node]:
     Returns:
         A list of Node.
     """
+    peers = list()
     try:
-        cmd = [f'{binpath}/mrtool',
-               'describe',
-               f'--cluster-id={cluster_id}',
-               f'--address={seed_addr}']
+        cmd = [
+            f"{binpath}/mrtool",
+            "describe",
+            f"--cluster-id={cluster_id}",
+            f"--address={seed_addr}",
+        ]
+        logger.info(f"fetch_peers: {cmd}")
         out = subprocess.check_output(cmd)
         cluster_info = json.loads(out)
-        members = cluster_info.get('members', dict())
-        peers = list()
+        members = cluster_info.get("members", dict())
         for node_id, member in members.items():
-            peer = member.get('peer', '')
-            endpoint = member.get('endpoint', '')
-            learner = member.get('learner', False)
+            peer = member.get("peer", "")
+            endpoint = member.get("endpoint", "")
+            learner = member.get("learner", False)
             if len(peer) == 0 or len(endpoint) == 0 or learner:
                 continue
             peers.append(Node(peer, endpoint))
-        for peer in peers:
-            logger.info(f'peer: raft_addr={peer.raft_address}\trpc_addr={peer.rpc_address}')
-        return peers
     except subprocess.CalledProcessError:
-        return list()
+        logger.info("could not fetch peers")
 
-
-def get_command(restart: bool, peers: list[Node]) -> list[str]:
-    cluster_id = get_cluster_id()
-    replication_factor = get_replication_factor()
-    raft_address = get_raft_address()
-    rpc_address = get_rpc_address()
-    raft_dir = get_raft_dir()
-    log_dir = get_log_dir()
-
-    cmd = [
-        f'{binpath}/vmr',
-        'start',
-        f'--cluster-id={cluster_id}',
-        f'--raft-address={raft_address}',
-        f'--bind={rpc_address}',
-        f'--log-rep-factor={replication_factor}',
-        f'--raft-dir={raft_dir}',
-        f'--log-dir={log_dir}'
-    ]
-    if len(peers) > 0:
-        cmd.append('--join=true')
-    if restart:
-        return cmd
     for peer in peers:
-        cmd.append(f'--peers={peer.raft_address}')
-    return cmd
+        logger.info(
+            f"peer: raft_addr={peer.raft_address}\trpc_addr={peer.rpc_address}"
+        )
+    return peers
 
 
-def main():
-    limits.set_limits()
-    prepare_vmr_home()
+def truncate(path: str) -> None:
+    """Truncate directory.
+
+    Args:
+        path: directory path
+    """
+    shutil.rmtree(path, ignore_errors=True)
+    os.makedirs(path, exist_ok=True)
+
+
+def start(
+    cluster_id: int,
+    raft_addr: str,
+    listen: str,
+    seed: str,
+    admin: str,
+    replication_factor: int,
+    add_after_seconds: int,
+    retry_interval_seconds: int,
+    raft_dir: str,
+    log_dir: str = None,
+) -> None:
+    """Start a metadata repository node.
+
+    Args:
+        cluster_id:Cluster ID.
+        raft_addr: An address to use RAFT protocol.
+        listen: An address to use RPC communications.
+        seed: Seed metadata repository address. For example, virtual IP for the
+          metadata repositories can be used.
+        admin: An address to connect to admin server.
+        replication_factor: Replication factor.
+        add_after_seconds: Add this node to the metadata repository cluster
+          after this number of seconds.
+        retry_interval_seconds: Retry interval seconds.
+        raft_dir: Path to the RAFT directory.
+        log_dir: Path to the log directory.
+    """
+    os.makedirs(raft_dir, exist_ok=True)
+    os.makedirs(log_dir, exist_ok=True)
 
     # If the join is true, the old write-ahead-log will be re-used.
-    restart = exists_wal(get_raft_dir())
+    restart = exists_wal(raft_dir)
     if restart:
-        logger.info(f'WAL already exists. (restart={restart})')
+        logger.info(f"WAL already exists. (restart={restart})")
     # if restart:
     #     # FIXME: Removing old WAL is only for testing.
     #     try:
@@ -235,24 +184,83 @@ def main():
     #     restart = False
 
     killer = Killer()
+    ok = False
     while not killer.kill_now:
-        if procutil.check_liveness('vmr'):
-            time.sleep(RETRY_INTERVAL_SEC)
+        if ok and procutil.check_liveness(APP_NAME):
+            time.sleep(retry_interval_seconds)
             continue
 
-        procutil.kill('vmr')
-        cluster_id = get_cluster_id()
-        seed_address = get_seed_address()
-        peers = fetch_peers(cluster_id, seed_address)
-        cmd = get_command(restart, peers)
-        logger.info(f'running metadata repository: {cmd}')
+        procutil.kill(APP_NAME)
+
+        peers = fetch_peers(cluster_id, seed)
+
+        cmd = [
+            f"{binpath}/vmr",
+            "start",
+            f"--cluster-id={cluster_id}",
+            f"--raft-address={raft_addr}",
+            f"--bind={listen}",
+            f"--log-rep-factor={replication_factor}",
+            f"--raft-dir={raft_dir}",
+            f"--log-dir={log_dir}",
+        ]
+        if len(peers) > 0 or restart:
+            cmd.append("--join=true")
+
+        if not restart:
+            for peer in peers:
+                cmd.append(f"--peers={peer.raft_address}")
+
+        logger.info(f"running metadata repository: {cmd}")
         subprocess.Popen(cmd)
+
+        time.sleep(add_after_seconds)
         if not restart and len(peers) > 0:
-            if not add_raft_peer():
-                logger.exception('could not add mr')
-        time.sleep(RETRY_INTERVAL_SEC)
-    procutil.stop('vmr')
+            added = add_raft_peer(admin, raft_addr, listen)
+            if not added:
+                procutil.kill(APP_NAME)
+                truncate(raft_dir)
+                logger.error("could not add mr")
+        ok = True
+        time.sleep(retry_interval_seconds)
+    procutil.stop(APP_NAME)
 
 
-if __name__ == '__main__':
+def main() -> None:
+    addr = socket.gethostbyname(socket.gethostname())
+    parser = argparse.ArgumentParser(description=APP_NAME)
+    parser.add_argument("--cluster-id", default=DEFAULT_CLUSTER_ID, type=int)
+    parser.add_argument("--raft-address",
+                        default=f"{addr}:{DEFAULT_RAFT_PORT}")
+    parser.add_argument("--listen", default=f"0.0.0.0:{DEFAULT_RPC_PORT}")
+    parser.add_argument("--seed", required=True)
+    parser.add_argument("--admin", required=True)
+    parser.add_argument("--replication-factor", required=True, type=int)
+    parser.add_argument("--raft-dir")
+    parser.add_argument("--log-dir")
+    parser.add_argument("--add-after-seconds",
+                        default=DEFAULT_ADD_AFTER_SECONDS,
+                        type=int)
+    parser.add_argument("--retry-interval-seconds",
+                        default=DEFAULT_RETRY_INTERVAL_SECONDS,
+                        type=int)
+    args = parser.parse_args()
+    logger.info(args)
+
+    limits.set_limits()
+    start(
+        args.cluster_id,
+        args.raft_address,
+        args.listen,
+        args.seed,
+        args.admin,
+        args.replication_factor,
+        args.add_after_seconds,
+        args.retry_interval_seconds,
+        args.raft_dir,
+        args.log_dir,
+    )
+
+
+if __name__ == "__main__":
     main()